+++ src/Files.App/Commands/Manager/CommandManager.cs
@@ -160,12 +160,6 @@ namespace Files.App.Commands
 		public IRichCommand GitFetch => commands[CommandCodes.GitFetch];
 		public IRichCommand GitPull => commands[CommandCodes.GitPull];
 
-		private IRichCommand _ModifiableDeleteItem;
-		public IRichCommand ModifiableDeleteItem => _ModifiableDeleteItem;
-
-		private IRichCommand _ModifiablePasteItem;
-		public IRichCommand ModifiablePasteItem => _ModifiablePasteItem;
-
 		public CommandManager()
 		{
 			commands = CreateActions()
@@ -173,14 +167,6 @@ namespace Files.App.Commands
 				.Cast<IRichCommand>()
 				.Append(new NoneCommand())
 				.ToImmutableDictionary(command => command.Code);
-
-			_ModifiableDeleteItem = new ModifiableCommand(commands[CommandCodes.DeleteItem], new() {
-				{ KeyModifiers.Shift,  commands[CommandCodes.DeleteItemPermanently] }
-			});
-
-			_ModifiablePasteItem = new ModifiableCommand(commands[CommandCodes.PasteItem], new() {
-				{ KeyModifiers.Shift,  commands[CommandCodes.PasteItemToSelection] }
-			});
 
 			settings.PropertyChanged += Settings_PropertyChanged;
 			UpdateHotKeys();
@@ -376,48 +362,8 @@ namespace Files.App.Commands
 				UpdateHotKeys();
 		}
 
-		[DebuggerDisplay("Command None")]
-		private class NoneCommand : IRichCommand
-		{
-			public event EventHandler? CanExecuteChanged { add {} remove {} }
-			public event PropertyChangingEventHandler? PropertyChanging { add {} remove {} }
-			public event PropertyChangedEventHandler? PropertyChanged { add {} remove {} }
-
-			public CommandCodes Code => CommandCodes.None;
-
-			public string Label => string.Empty;
-			public string LabelWithHotKey => string.Empty;
-			public string AutomationName => string.Empty;
-
-			public string Description => string.Empty;
-
-			public RichGlyph Glyph => RichGlyph.None;
-			public object? Icon => null;
-			public FontIcon? FontIcon => null;
-			public Style? OpacityStyle => null;
-
-			public bool IsCustomHotKeys => false;
-			public string? HotKeyText => null;
-			public HotKeyCollection HotKeys
-			{
-				get => HotKeyCollection.Empty;
-				set => throw new InvalidOperationException("This command is readonly.");
-			}
-
-			public bool IsToggle => false;
-			public bool IsOn { get => false; set {} }
-			public bool IsExecutable => false;
-
-			public bool CanExecute(object? parameter) => false;
-			public void Execute(object? parameter) {}
-			public Task ExecuteAsync() => Task.CompletedTask;
-			public void ExecuteTapped(object sender, TappedRoutedEventArgs e) {}
-
-			public void ResetHotKeys() {}
-		}
-
 		[DebuggerDisplay("Command {Code}")]
-		private class ActionCommand : ObservableObject, IRichCommand
+		internal class ActionCommand : ObservableObject, IRichCommand
 		{
 			public event EventHandler? CanExecuteChanged;
 
@@ -540,113 +486,6 @@ namespace Files.App.Commands
 					OnPropertyChanged(nameof(LabelWithHotKey));
 				}
 			}
-
-			private void Action_PropertyChanging(object? sender, PropertyChangingEventArgs e)
-			{
-				switch (e.PropertyName)
-				{
-					case nameof(IAction.Label):
-						OnPropertyChanging(nameof(Label));
-						OnPropertyChanging(nameof(LabelWithHotKey));
-						OnPropertyChanging(nameof(AutomationName));
-						break;
-					case nameof(IToggleAction.IsOn) when IsToggle:
-						OnPropertyChanging(nameof(IsOn));
-						break;
-					case nameof(IAction.IsExecutable):
-						OnPropertyChanging(nameof(IsExecutable));
-						break;
-				}
-			}
-			private void Action_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-			{
-				switch (e.PropertyName)
-				{
-					case nameof(IAction.Label):
-						OnPropertyChanged(nameof(Label));
-						OnPropertyChanged(nameof(LabelWithHotKey));
-						OnPropertyChanged(nameof(AutomationName));
-						break;
-					case nameof(IToggleAction.IsOn) when IsToggle:
-						OnPropertyChanged(nameof(IsOn));
-						break;
-					case nameof(IAction.IsExecutable):
-						OnPropertyChanged(nameof(IsExecutable));
-						CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-						break;
-				}
-			}
-		}
-
-		[DebuggerDisplay("Command {Code} (Modifiable)")]
-		private class ModifiableCommand : ObservableObject, IRichCommand
-		{
-			public event EventHandler? CanExecuteChanged;
-
-			private IRichCommand BaseCommand;
-			private ImmutableDictionary<KeyModifiers, IRichCommand> ModifiedCommands;
-
-			public CommandCodes Code => BaseCommand.Code;
-
-			public string Label => BaseCommand.Label;
-			public string LabelWithHotKey => BaseCommand.LabelWithHotKey;
-			public string AutomationName => BaseCommand.AutomationName;
-
-			public string Description => BaseCommand.Description;
-
-			public RichGlyph Glyph => BaseCommand.Glyph;
-			public object? Icon => BaseCommand.Icon;
-			public FontIcon? FontIcon => BaseCommand.FontIcon;
-			public Style? OpacityStyle => BaseCommand.OpacityStyle;
-
-			public bool IsCustomHotKeys => BaseCommand.IsCustomHotKeys;
-			public string? HotKeyText => BaseCommand.HotKeyText;
-
-			public HotKeyCollection HotKeys
-			{
-				get => BaseCommand.HotKeys;
-				set => BaseCommand.HotKeys = value;
-			}
-
-			public bool IsToggle => BaseCommand.IsToggle;
-
-			public bool IsOn
-			{
-				get => BaseCommand.IsOn;
-				set => BaseCommand.IsOn = value;
-			}
-
-			public bool IsExecutable => BaseCommand.IsExecutable;
-
-			public ModifiableCommand(IRichCommand baseCommand, Dictionary<KeyModifiers, IRichCommand> modifiedCommands)
-			{
-				BaseCommand = baseCommand;
-				ModifiedCommands = modifiedCommands.ToImmutableDictionary();
-
-				if (baseCommand is ActionCommand actionCommand)
-				{
-					if (actionCommand.Action is INotifyPropertyChanging notifyPropertyChanging)
-						notifyPropertyChanging.PropertyChanging += Action_PropertyChanging;
-					if (actionCommand.Action is INotifyPropertyChanged notifyPropertyChanged)
-						notifyPropertyChanged.PropertyChanged += Action_PropertyChanged;
-				}
-			}
-
-			public bool CanExecute(object? parameter) => BaseCommand.CanExecute(parameter);
-			public async void Execute(object? parameter) => await ExecuteAsync();
-
-			public Task ExecuteAsync()
-			{
-				if (ModifiedCommands.TryGetValue(HotKeyHelpers.GetCurrentKeyModifiers(), out var modifiedCommand) &&
-					modifiedCommand.IsExecutable)
-					return modifiedCommand.ExecuteAsync();
-				else
-					return BaseCommand.ExecuteAsync();
-			}
-
-			public async void ExecuteTapped(object sender, TappedRoutedEventArgs e) => await ExecuteAsync();
-
-			public void ResetHotKeys() => BaseCommand.ResetHotKeys();
 
 			private void Action_PropertyChanging(object? sender, PropertyChangingEventArgs e)
 			{
