+++ src/Files.App/Commands/Manager/ICommandManager.cs
@@ -162,8 +162,5 @@ namespace Files.App.Commands
 
 		IRichCommand GitFetch { get; }
 		IRichCommand GitPull { get; }
-
-		IRichCommand ModifiableDeleteItem { get; }
-		IRichCommand ModifiablePasteItem { get; }
 	}
 }
